requirements.txt
django==3.2.25
djangorestframework==3.15.1
django-cors-headers==4.1.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==0.21.1
bcrypt==4.2.1
PyJWT==2.8.0
openpyxl==3.1.3
pymysql==1.1.0
